+++ Code/Test/UI/TestHealthBar.cs
@@ -19,8 +19,8 @@ public class TestHealthBar
     [TestCase]
     public void CorrectlyInitializeHealthValues()
     {
-        int maxHealth = hb.maxHP;
-        int currentHealth = hb.currentHP;
+        float maxHealth = hb.maxHP;
+        float currentHealth = hb.currentHP;
 
         Assertions.AssertFloat(maxHealth).Equals(hb.maxHP);
         Assertions.AssertFloat(maxHealth).Equals(hb.MaxValue);
@@ -38,7 +38,7 @@ public class TestHealthBar
 
     public void OnHealthChangedUpdatesHealthValues()
     {
-        int newHealth = 10;
+        float newHealth = 10;
         hb.OnHealthChanged(newHealth);
 
         Assertions.AssertFloat(newHealth).Equals(hb.currentHP);
@@ -52,7 +52,7 @@ public class TestHealthBar
     [TestCase]
     public void OnHealthChangedStartsRedHealthTimer()
     {
-        int newHealth = 10;
+        float newHealth = 10;
         hb.OnHealthChanged(newHealth);
 
         Assertions.AssertBool(hb.redHealthTimer.IsStopped()).IsFalse();
@@ -61,7 +61,7 @@ public class TestHealthBar
 
     public void OnMaxHealthChangedUpdatesMaxHealthValues()
     {
-        int newMaxHealth = 100;
+        float newMaxHealth = 100;
         hb.OnMaxHealthChanged(newMaxHealth);
 
         Assertions.AssertFloat(newMaxHealth).Equals(hb.maxHP);
